requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==2.3.0
sqlalchemy==2.0.18
asyncpg==0.28.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==4.5.5
sympy==1.12
numpy==1.26.1
scipy==1.11.1
python-multipart==0.0.6
prometheus-client==0.16.0
httpx==0.24.1
python-dotenv==1.0.0
pytest==7.3.2
gunicorn==22.1.0
uvloop==0.19.0
